fix logging warnings

+++ src/hostapi/coreaudio/pa_mac_core.c
@@ -373,7 +373,7 @@ static PaError gatherDeviceInfo(PaMacAUHAL *auhalHostApi)
     {
        int i;
        for( i=0; i<auhalHostApi->devCount; ++i )
-          printf( "Device %d\t: %ld\n", i, auhalHostApi->devIds[i] );
+          printf( "Device %d\t: %ld\n", i, (long)auhalHostApi->devIds[i] );
     }
 #endif
 
@@ -420,8 +420,8 @@ static PaError gatherDeviceInfo(PaMacAUHAL *auhalHostApi)
        }
     }   
 
-    VDBUG( ( "Default in : %ld\n", auhalHostApi->defaultIn  ) );
-    VDBUG( ( "Default out: %ld\n", auhalHostApi->defaultOut ) );
+    VDBUG( ( "Default in : %ld\n", (long)auhalHostApi->defaultIn  ) );
+    VDBUG( ( "Default out: %ld\n", (long)auhalHostApi->defaultOut ) );
 
     return paNoError;
 }
@@ -1606,15 +1606,15 @@ static PaError OpenAndSetupOneAudioUnit(
 
     if( inStreamParams && outStreamParams )
     {
-        VDBUG( ("Opened device %ld for input and output.\n", *audioDevice ) );
+        VDBUG( ("Opened device %ld for input and output.\n", (long)*audioDevice ) );
     }
     else if( inStreamParams )
     {
-        VDBUG( ("Opened device %ld for input.\n", *audioDevice ) );
+        VDBUG( ("Opened device %ld for input.\n", (long)*audioDevice ) );
     }
     else if( outStreamParams )
     {
-        VDBUG( ("Opened device %ld for output.\n", *audioDevice ) );
+        VDBUG( ("Opened device %ld for output.\n", (long)*audioDevice ) );
     }
     return paNoError;
 #undef ERR_WRAP
@@ -1676,7 +1676,7 @@ static UInt32 CalculateOptimalBufferSize( PaMacAUHAL *auhalHostApi,
         
     }else{
     	VDBUG( ("Block Size unspecified. Based on Latency, the user wants a Block Size near: %ld.\n",
-            resultBufferSizeFrames ) );
+            (long)resultBufferSizeFrames ) );
     }
     
     // Clip to the capabilities of the device.
@@ -1691,7 +1691,7 @@ static UInt32 CalculateOptimalBufferSize( PaMacAUHAL *auhalHostApi,
         ClipToDeviceBufferSize( auhalHostApi->devIds[outputParameters->device],
                                false, resultBufferSizeFrames, &resultBufferSizeFrames );
     }
-    VDBUG(("After querying hardware, setting block size to %ld.\n", resultBufferSizeFrames));
+    VDBUG(("After querying hardware, setting block size to %ld.\n", (long)resultBufferSizeFrames));
 
     return resultBufferSizeFrames;
 }
